move TokenEmulator to TokenEmulator namespace
