+++ lib/PhpParser/Lexer/TokenEmulator/FnTokenEmulator.php
@@ -1,21 +1,22 @@
 <?php declare(strict_types=1);
 
-namespace PhpParser\Lexer;
+namespace PhpParser\Lexer\TokenEmulator;
 
+use PhpParser\Lexer\Emulative;
 use PhpParser\Parser\Tokens;
 
-final class CoaleseEqualTokenEmulator implements TokenEmulatorInterface
+final class FnTokenEmulator implements TokenEmulatorInterface
 {
-    const T_COALESCE_EQUAL = 1007;
+    const T_FN = 1008;
 
     public function getTokenId(): int
     {
-        return self::T_COALESCE_EQUAL;
+        return self::T_FN;
     }
 
     public function getParserTokenId(): int
     {
-        return Tokens::T_COALESCE_EQUAL;
+        return Tokens::T_FN;
     }
 
     public function isEmulationNeeded(string $code) : bool
@@ -25,26 +26,20 @@ final class CoaleseEqualTokenEmulator implements TokenEmulatorInterface
             return false;
         }
 
-        return strpos($code, '??=') !== false;
+        return strpos($code, 'fn') !== false;
     }
 
     public function emulate(string $code, array $tokens) : array
     {
         // We need to manually iterate and manage a count because we'll change
         // the tokens array on the way
-        $line = 1;
         for ($i = 0, $c = count($tokens); $i < $c; ++$i) {
-            if (isset($tokens[$i + 1])) {
-                if ($tokens[$i][0] === T_COALESCE && $tokens[$i + 1] === '=') {
-                    array_splice($tokens, $i, 2, [
-                        [self::T_COALESCE_EQUAL, '??=', $line]
-                    ]);
-                    $c--;
+            if ($tokens[$i][0] === T_STRING && $tokens[$i][1] === 'fn') {
+                if (isset($tokens[$i - 1]) && $tokens[$i - 1][0] === T_OBJECT_OPERATOR) {
                     continue;
                 }
-            }
-            if (\is_array($tokens[$i])) {
-                $line += substr_count($tokens[$i][1], "\n");
+
+                $tokens[$i][0] = self::T_FN;
             }
         }
