+++ lib/PhpParser/Lexer/TokenEmulator/TokenEmulatorInterface.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace PhpParser\Lexer;
+namespace PhpParser\Lexer\TokenEmulator;
 
 interface TokenEmulatorInterface
 {
